+++ source_gen/source_gen/test/src/unformatted_code_generator.dart
@@ -1,0 +1,26 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:async';
+
+import 'package:analyzer/dart/element/element.dart';
+import 'package:source_gen/source_gen.dart';
+
+/// Generates a single-line of unformatted code.
+class UnformattedCodeGenerator extends Generator {
+  const UnformattedCodeGenerator();
+
+  @override
+  Future<String> generate(Element element, _) async {
+    return unformattedCode;
+  }
+
+  static const formattedCode = '''
+void hello() => print('hello');
+''';
+
+  static const unformattedCode = '''
+void hello ()=>  print('hello');
+''';
+}